+++ apps/ui/CHANGELOG.md
@@ -1,11 +1,11 @@
-# @dojocho/ui
+# @dojofoo/lesson-ui
 
 ## 0.0.3
 
 ### Patch Changes
 
 - Updated dependencies [d87d2e2]
-  - @dojocho/config@0.0.5
+  - @dojofoo/config@0.0.5
 
 ## 0.0.2