+++ dojos/pydantic-agents/CHANGELOG.md
@@ -1,11 +1,11 @@
-# @dojocho/pydantic-agents
+# @dojofoo/pydantic-agents
 
 ## 0.0.4
 
 ### Patch Changes
 
 - Updated dependencies [d87d2e2]
-  - @dojocho/config@0.0.5
+  - @dojofoo/config@0.0.5
 
 ## 0.0.3